backend-studenttalent/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow
psycopg2-binary
dj-database-url
python-dotenv
requests
gunicorn
whitenoise
